docs/requirements.txt
sphinx>=4
sphinx_copybutton
pydata_sphinx_theme
pandas
ruamel.yaml
myst_parser
